Revert "Initial Cut of Claim [Pension/Burial] Uploads and Processing"

+++ lib/workflow/task/common/datestamp_pdf_task.rb
@@ -1,22 +1,15 @@
 # frozen_string_literal: true
-require 'workflow/task/shrine_file/base'
-
-require 'fileutils'
-module Workflow::Task::Shared
+module Workflow::Task::Common
   class DatestampPdfTask < Workflow::Task::ShrineFile::Base
     def run(settings)
-      FileUtils.mkdir_p(Rails.root.join('tmp', 'pdfs'))
       in_path = @file.download.path
+      FileUtils.mkdir_p Rails.root.join('tmp', 'pdfs')
       stamp_path = Rails.root.join('tmp', 'pdfs', "#{SecureRandom.uuid}.pdf")
       generate_stamp(stamp_path, settings[:text], settings[:x], settings[:y])
       out_path = stamp(in_path, stamp_path)
       update_file(io: File.open(out_path))
     ensure
-      File.delete(stamp_path) if stamp_path && File.exist?(stamp_path)
-      if out_path && File.exist?(out_path)
-        File.delete(out_path)
-        FileUtils.rmdir(File.dirname(out_path))
-      end
+      File.delete(out_path) if !out_path.nil? && File.exist?(out_path)
     end
 
     private
@@ -33,19 +26,17 @@ module Workflow::Task::Shared
     end
 
     def stamp(file_path, stamp_path)
-      out_dir = Rails.root.join('tmp', 'pdfs', SecureRandom.uuid)
-      FileUtils.mkdir_p(out_dir)
-      out_path = File.join(out_dir, @file.original_filename)
+      out_path = Rails.root.join('tmp', 'pdfs', "#{SecureRandom.uuid}.pdf")
       stamp = CombinePDF.load(stamp_path).pages[0]
       original = CombinePDF.load(file_path)
       original.pages.each { |page| page << stamp }
       original.save out_path
       out_path
     rescue => e
-      File.delete(out_path) if out_path && File.exist?(out_path)
-      FileUtils.rmdir(out_dir) if out_dir && Dir.exist?(out_dir)
       Rails.logger.error "Failed to datestamp PDF file: #{e.message}"
       raise
+    ensure
+      File.delete(stamp_path) if File.exist?(stamp_path)
     end
   end
 end
